# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SupervIA - AI Service API",
    "version": "1.0.0",
    "description": "API du service d'intelligence artificielle de SupervIA. Il expose deux agents principaux : le **Dashboard Agent** pour la génération de tableaux de bord, et l'**Ops Assistant** pour le diagnostic et l'assistance opérationnelle via un chat interactif.",
    "contact": {
      "name": "SupervIA Team",
      "url": "https://github.com/zkerkeb-class/service-ia-Tadayoshi123"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3004",
      "description": "Serveur de développement local"
    },
    {
      "url": "http://localhost:8080/api/v1/ai",
      "description": "Via la passerelle API Nginx"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token d'accès JWT (RS256) pour l'authentification des utilisateurs finaux."
      }
    },
    "schemas": {
...
      "CacheStats": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "integer"
          },
          "hits": {
            "type": "integer"
          },
          "misses": {
            "type": "integer"
          },
          "hitRate": {
            "type": "number",
            "format": "float"
          },
          "memoryUsage": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/api/cache/stats": {
      "get": {
        "summary": "Obtenir les statistiques du cache Redis",
        "tags": [
          "Cache"
        ],
        "security": [
          {
            "UserAuth": []
          }
        ],
        "description": "Retourne des statistiques d'utilisation du cache Redis, comme le nombre de clés, le taux de succès (hit rate) et l'utilisation de la mémoire. Nécessite des droits d'administrateur.",
        "responses": {
          "200": {
            "description": "Statistiques du cache récupérées avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheStats"
                }
              }
            }
          },
          "401": {
            "description": "Non autorisé."
          },
          "403": {
            "description": "Permissions insuffisantes."
          }
        }
      }
    },
    "/api/cache/flush": {
      "delete": {
        "summary": "Vider complètement le cache Redis",
        "tags": [
          "Cache"
        ],
        "security": [
          {
            "UserAuth": []
          }
        ],
        "description": "Supprime toutes les clés du cache Redis pour la base de données configurée. Opération destructive. Nécessite des droits d'administrateur.",
        "responses": {
          "200": {
            "description": "Cache vidé avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cache flushed successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autorisé."
          },
          "403": {
            "description": "Permissions insuffisantes."
          }
        }
      }
    },
    "/api/cache/keys": {
      "get": {
        "summary": "Lister les clés du cache",
        "tags": [
          "Cache"
        ],
        "security": [
          {
            "UserAuth": []
          }
        ],
        "description": "Récupère une liste paginée de toutes les clés actuellement dans le cache Redis. Nécessite des droits d'administrateur.",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Le curseur pour la pagination (obtenu depuis la réponse précédente)."
          },
          {
            "in": "query",
            "name": "count",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Le nombre de clés à retourner par page."
          }
        ],
        "responses": {
          "200": {
            "description": "Liste de clés récupérée."
          },
          "401": {
            "description": "Non autorisé."
          },
          "403": {
            "description": "Permissions insuffisantes."
          }
        }
      }
    },
    "/api/cache/keys/{key}": {
      "delete": {
        "summary": "Supprimer une clé spécifique du cache",
        "tags": [
          "Cache"
        ],
        "security": [
          {
            "UserAuth": []
          }
        ],
        "description": "Supprime une clé unique du cache Redis. Nécessite des droits d'administrateur.",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "La clé à supprimer (doit être encodée en URL si elle contient des caractères spéciaux)."
          }
        ],
        "responses": {
          "200": {
            "description": "Clé supprimée avec succès."
          },
          "401": {
            "description": "Non autorisé."
          },
          "403": {
            "description": "Permissions insuffisantes."
          },
          "404": {
            "description": "Clé non trouvée."
          }
        }
      }
    },
...
  }
}